+++ README.md
@@ -57,3 +57,7 @@ To build all apps and packages, run the following command:
 ```
 pnpm run build
 ```
+DEV PREPARING (WHEN MOVE TO NEW DEVICE)
+npm i
+npx prisma generate
+Not sure, but |=> npx prisma migrate dev --name init
