--- apps/frontend/prisma/schema.prisma
@@ -1,65 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "mongodb"
-  url      = env("DATABASE_URL")
-}
-
-model Book {
-    id String @id @default(auto()) @map("_id") @db.ObjectId
-    name String
-    ISBN String
-    authorId String @db.ObjectId
-    publisherId String @db.ObjectId
-    description String
-    genresIds String[] @db.ObjectId
-    reviewsIds String[] @db.ObjectId
-    rating Float
-    coverImageURL String
-    keywords String[]
-    publishDate DateTime
-    fileURL String
-    createdAt DateTime
-    updatedAt DateTime
-}
-
-model User {
-    id String @id @default(auto()) @map("_id") @db.ObjectId
-    username String
-    birthday DateTime
-    role UserRole
-    // Что-то с жанрами
-    // interestsIds String[] 
-    // readedBooksIds String[] @db.ObjectId
-    // favoritesIds String[] @db.ObjectId
-    email String
-    avatarURL String
-    password String
-    createdAt DateTime
-    updatedAt DateTime
-}
-
-model Author {
-    id String @id @default(auto()) @map("_id") @db.ObjectId
-    name String
-    surname String
-    description String
-    writtenBooksIds String[] @db.ObjectId
-    birthday DateTime
-    dayOfDeath DateTime
-    secondName String?
-}
-
-enum UserRole {
-    USER
-    AUTHOR
-    MODERATOR
-    ADMINISTRATOR
-}
-
-
