+++ package.json
@@ -12,6 +12,7 @@
     "lint:fix": "eslint --fix"
   },
   "devDependencies": {
+    "@types/node": "^18.15.11",
     "@vueuse/core": "^9.13.0",
     "@vueuse/nuxt": "^9.13.0",
     "nuxt": "3.3.3",
