+++ prisma/schema.prisma
@@ -1,0 +1,57 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "sqlite"
+  url      = env("DATABASE_URL")
+}
+
+model Book {
+    id String @id @default(cuid())
+    name String
+    ISBN String
+}
+//     authorId String
+//     publisherId String 
+//     description String
+//     genresIds String[]
+//     reviewsIds String[]
+//     rating Float
+//     coverImageURL String
+//     keywords String[]
+//     publishDate DateTime
+//     fileURL String
+//     createdAt DateTime
+//     updatedAt DateTime
+// }
+
+// model User {
+//     id String @id @default(cuid())
+//     username String
+//     birthday DateTime
+//     role String
+//     // Что-то с жанрами
+//     interestsIds String[]
+//     readedBooksIds String[]
+//     favoritesIds String[]
+//     email String
+//     avatarURL String
+//     password String
+//     createdAt DateTime
+//     updatedAt DateTime
+// }
+
+// model Author {
+//     id String @id @default(cuid())
+//     name String
+//     surname String
+//     description String
+//     writtenBooksIds String[] 
+//     birthday DateTime
+//     dayOfDeath DateTime
+//     secondName String?
+// }
